fiddling the detection algorithm

+++ embedded/compteur/histogram.h
@@ -34,8 +34,8 @@ namespace histogram {
         usize size() const;
         //! number of update() calls - 1 (first sets min/max)
         u32 count() const;
-        u16 minimum() const;
-        u16 maximum() const;
+        float minimum() const;
+        float maximum() const;
         /*! 
 		  arghigh(0) = maximum()
 		  arghigh(100) = minimum()
@@ -49,7 +49,7 @@ namespace histogram {
             *L = sizeof(m_packed);
             return &m_packed;
         }
-        u16 min(u16 index) const;
+        float min(u16 index) const;
         u32 count(u16 index) const;
         u16 argmax(u16 m, u16 M) const;
         u16 argmin(u16 m, u16 M) const;
